requirements.txt
Flask==2.2.2
Flask_Cors==4.0.0
openai==1.9.0
pymongo==4.6.1
protobuf==4.25.2
Werkzeug==2.2.2
pinecone-client